Create the useradd cli.
There isn't a great way to create a user from within the ruby standard
library, so I opted to just shell-out to create the user. This is
assuming the sysadmin is running this file directly on the box itself.
Using net/ssh to do this remotely is certainly possible if requested.